Code clean up
- Remove weight files
- Different naming
