+++ parc/layer.py
@@ -16,8 +16,8 @@ def resnet_unit(feat_dim, kernel_size, x_in, padding="CONSTANT"):
     Return:
                 - (tensor): output tensor of residual unit.
     """
-    to_pad = (kernel_size - 1) // 2
-    pad_instruct = tf.constant([[0, 0], [to_pad, to_pad], [to_pad, to_pad], [0, 0]])
+    pad_size = (kernel_size - 1) // 2
+    pad_instruct = tf.constant([[0, 0], [pad_size, pad_size], [pad_size, pad_size], [0, 0]])
     x = tf.pad(x_in, pad_instruct, padding)
     x = Conv2D(feat_dim, kernel_size, padding="valid")(x)
     x = ReLU()(x)
@@ -39,8 +39,8 @@ def resnet_block(x_in, feat_dim, kernel_size, reps, pooling=True, padding="CONST
     Return:
                 - (tensor): output of the resnet block
     """
-    to_pad = (kernel_size - 1) // 2
-    pad_instruct = tf.constant([[0, 0], [to_pad, to_pad], [to_pad, to_pad], [0, 0]])
+    pad_size = (kernel_size - 1) // 2
+    pad_instruct = tf.constant([[0, 0], [pad_size, pad_size], [pad_size, pad_size], [0, 0]])
     x = tf.pad(x_in, pad_instruct, padding)
     x = Conv2D(feat_dim, kernel_size, padding="valid")(x)
     x = ReLU()(x)
@@ -66,8 +66,8 @@ def conv_unit(feat_dim, kernel_size, x_in, padding="CONSTANT"):
     Return:
                 - (tensor): output of the conv unit
     """
-    to_pad = (kernel_size - 1) // 2
-    pad_instruct = tf.constant([[0, 0], [to_pad, to_pad], [to_pad, to_pad], [0, 0]])
+    pad_size = (kernel_size - 1) // 2
+    pad_instruct = tf.constant([[0, 0], [pad_size, pad_size], [pad_size, pad_size], [0, 0]])
     x = tf.pad(x_in, pad_instruct, padding)
     x = Conv2D(feat_dim, kernel_size, activation=LeakyReLU(0.2), padding="valid")(x)
     x = Conv2D(feat_dim, 1, activation=LeakyReLU(0.2), padding="valid")(x)
@@ -136,21 +136,20 @@ def spade_generator_unit(x, mask, feats_out, kernel, upsampling=True, padding="C
                                         |                               |
                                          ---- (SPADE + Relu + Conv) ----
     """
-    to_pad = (kernel - 1) // 2
-    pad_instruct = tf.constant([[0, 0], [to_pad, to_pad], [to_pad, to_pad], [0, 0]])
+    pad_size = (kernel - 1) // 2
+    pad_instruct = tf.constant([[0, 0], [pad_size, pad_size], [pad_size, pad_size], [0, 0]])
     x = GaussianNoise(0.05)(x)
     # Residual SPADE & conv
-    spade1 = base.SPADE(feats_out)(x, mask)
+    spade1 = SPADE(feats_out)(x, mask)
     relu1 = LeakyReLU(0.2)(spade1)
-
     relu1 = tf.pad(relu1, pad_instruct, padding)
     conv1 = Conv2D(feats_out, kernel, padding='valid')(relu1)
     spade2 = SPADE(feats_out, padding=padding)(conv1, mask)
     relu2 = LeakyReLU(0.2)(spade2)
     relu2 = tf.pad(relu2, pad_instruct, padding)
     conv2 = Conv2D(feats_out, kernel, padding='valid')(relu2)
-    # Skip
 
+    # Skip
     spade_skip = SPADE(feats_out, padding=padding)(x, mask)
     relu_skip = LeakyReLU(0.2)(spade_skip)
     relu_skip = tf.pad(relu_skip, pad_instruct, padding)
